requirements.txt
Django==4.1.1
djangorestframework==3.13.1
channels==3.0.5
Pillow==9.2.0